refactor(agents): split subagent registry machinery
